+++ SkillAPI/src/com/sucy/skill/command/admin/CmdExpPlayer.java
@@ -41,7 +41,7 @@ public class CmdExpPlayer implements ICommand {
             // Get the amount
             int amount = 0;
             try {
-                amount = Integer.parseInt(args[0]);
+                amount = (int)Double.parseDouble(args[0]);
             }
             catch (Exception ex) {
                 // Do nothing
